Use GraphQL on CI to get the list of modules
This adds a GH Action to fetch the list of modules under our org via GraphQL.

This list is stored as a comma-separated env-var in 'name@default-branch' format, and passed into the build.

We have to then filter this list of modules to remove those that are not part of the Micronaut release cycle.

If there is no env var, then we fall back to using the list of modules as before