Fix incorrectly detected cycle with test resources
The test resources server makes use of Micronaut Platform via the
Micronaut Gradle plugin, but it uses a snapshot version of the
plugin and platform. This will _not_ prevent from releasing, as
what we ship is a fat jar which contains all required dependencies,
and no dependency on the platform itself.