Add implicit dependency from gradle to platform
The Gradle plugin depends on Micronaut Platform, but it is not
visible as a runtime dependency, because it's a dependency which
is added at runtime _when the plugin is applied_. If we released
the Gradle plugin simply considering the current graph, then a
user applying the plugin would get a SNAPSHOT version of the
platform.

This commit updates the report tool to add a dependency so that
it's clearly visible.