Add basic search/filtering
I struggled to find the module of interest.
This PR adds a basic search filter to the page.

Just load the page, then type the module name and the list on the left will be filtered to match what you type.
Any non a-z character (such as Esc) clears the filter.

No UI, because I couldn't face it.